+++ src/catalib/cli/app.py
@@ -14,6 +14,7 @@ import typer
 from catalib import __version__, check_for_updates
 from catalib.cli.build_command import build
 from catalib.cli.init_command import init
+from catalib.cli.stubs_command import stubs
 from catalib.cli.watch_command import watch_command
 
 app = typer.Typer(
@@ -26,6 +27,7 @@ app = typer.Typer(
 app.command("build")(build)
 app.command("init")(init)
 app.command("watch")(watch_command)
+app.command("stubs")(stubs)
 
 
 @app.command("version")